+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -14,9 +14,9 @@ import org.apache.catalina.auth.HttpCookie;
 import org.apache.catalina.auth.Session;
 import org.apache.catalina.auth.SessionManager;
 import org.apache.catalina.http.VersionOfProtocol;
-import org.apache.catalina.io.FileReader;
-import org.apache.catalina.io.RequestParser;
-import org.apache.catalina.io.RequestReader;
+import org.apache.catalina.parser.RequestParser;
+import org.apache.catalina.reader.FileReader;
+import org.apache.catalina.reader.RequestReader;
 import org.apache.catalina.request.HttpMethod;
 import org.apache.catalina.request.HttpRequest;
 import org.apache.catalina.request.RequestBody;
